17-10-23 Crud Book

+++ 23-24/3A42/src/Controller/AuthorController.php
@@ -95,8 +95,14 @@ class AuthorController extends AbstractController
     {
         $author= $repository->find($id);
         $em= $managerRegistry->getManager();
-        $em->remove($author);
-        $em->flush();
+        if($author->getNbrBooks()==0){
+            $em->remove($author);
+            $em->flush();
+        }
+        else{
+            return new Response("Error");
+        }
+
         return $this->redirectToRoute("list_authors");
 
     }
